CHROMIUM: WAR: usb: dwc3: qcom: Put dwc3 device reference on failure
In the even that we successfully acquire a reference to the core dwc3
platform_device, but then fail to acquire the extcon of the qcom-dwc3
device we would leave a dangling reference, which means resources
associated with the dwc3 platform_device would never be released.

This consistently shows up as 4 entries in kmemleak when
dwc3_qcom_register_extcon() fails, so put the refcount of the acquired
dwc3 to ensure that the associated memory is released.

BUG=b:138962769
TEST=kmemleak should not show 4 entries when dwc3_qcom_register_extcon() fails.

Change-Id: I8b5ef90836a7a57487e59740404b8880a799ac33
Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
Signed-off-by: Khasim Syed Mohammed <khasim.mohammed@linaro.org>