arm64: dts: qcom: qcs404-evb: Move USB regulators out of SPMI block
The VBUS boost regulator is not located inside the PMS405 it is a discreet
chip toggled directly by GPIO3 as such we should not model it as a child of
the PMS405.

Move the location of the regulator declaration now to represent the
regulator as a standalone, not a subordinate of the PMS405.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>