usb: dwc3: qcom: dwc3_qcom_vbus_overrride_enable in OTG mode
Currently if the controller is in peripheral mode
dwc3_qcom_vbus_overrride_enable() is called. This patch makes that call
happen for OTG also.

Probably a bit of a hack looks like role-switch needs to be added to this
driver to switch between VBUS over-ride or not when extcon is not active.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>