clk: qcom: add gcc-msm8939: Fix bimc_ddr_clk_src rcgr base address
We copied directly from gcc-msm8916.c to gcc-msm8939.c. Looking into a
bandwidth issue on 8939 and checking the clocks as part of that it looks as
if the bmic_ddr_clk_src rcgr is wrong for 8939.

Documentation for 8939 shows "0x01832024 GCC_DDR_CMD_RCGR" its unclear if
this value is incorrect on msm8916 and should similarly be investigated.

Fixes: 2cbe54d0ae6b ("clk: qcom: add gcc-msm8939 driver by copying 8916")
Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>