clk: qcom: common: Add support for power-domain attachment
Right now we have a plethora of singleton power-domains which the core
automatically attaches to the clock however, when we need to deal with
multiple power-domains associated with one controller, the platform code
will no longer set this up for us.

Use dev_pm_domain_attach_list() to automatically hook the list of given
power-domains in the dtsi for the clock being registered in
qcom_cc_really_probe().

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>