clk: qcom: gdsc: Add pm_runtime hooks
The clock controller block which the GDSC is attached to needs to woken up
prior to trying to switch a GDSC on.

There is currently no established link between a GDSC and the power-domain
or power-domains of the clock the GDSC resides within, the platform code
just switches the power-domain on the GDSCs function fine.

The power-domains associated with clock controllers appear to get switched
on at probe time but then never to be switched off again until the
clock-controller exits, defacto meaning it never gets switched off.

When we have multiple power-domains associated with a clock controller we
are required to and have the opportunity to setup the link to the
power-domain for the clock controllers ourselves.

This type of setup doesn't just switch on it has to be requested to be
switched on by the pm runtime logic.

Add pm runtime logic to the gdsc_enable/gdsc_disable path taking care to
leave RETAIN_FF GDSCs with power-domains switched on.

This way we can power-up the power-domain or power-domains the clock
controller requires on the gdsc_enable() path.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>